+++ src/Ratchet/Wamp/Topic.php
@@ -66,7 +66,7 @@ class Topic implements \IteratorAggregate, \Countable
      */
     public function has(ConnectionInterface $conn)
     {
-        return $this->subscribers->contains($conn);
+        return $this->subscribers->offsetExists($conn);
     }
 
     /**
@@ -75,7 +75,7 @@ class Topic implements \IteratorAggregate, \Countable
      */
     public function add(ConnectionInterface $conn)
     {
-        $this->subscribers->attach($conn);
+        $this->subscribers->offsetSet($conn);
 
         return $this;
     }
@@ -86,8 +86,8 @@ class Topic implements \IteratorAggregate, \Countable
      */
     public function remove(ConnectionInterface $conn)
     {
-        if ($this->subscribers->contains($conn)) {
-            $this->subscribers->detach($conn);
+        if ($this->subscribers->offsetExists($conn)) {
+            $this->subscribers->offsetUnset($conn);
         }
 
         return $this;